requirements.txt
pylint
pytest
black
ipython
click
ipdb
openai
pytest-cov
pydantic
python-dotenv
pandas
qdrant-client
sentence-transformers
streamlit
geopy
fastapi
uvicorn[standard]